Adapters

+++ scrapers/mexico.py
@@ -70,21 +70,22 @@ class Mexico:
                 os.mkdir(self.downloads_dir)
 
         options = Options()
+        options.add_argument('--log-level=3')
         options.headless = True
         prefs = {"download.default_directory": self.downloads_dir}
         options.add_experimental_option("prefs", prefs)
 
         operating_system = platform.system()
-        chrome_driver = 'mac_chromedriver86'
+        chrome_driver = 'mac_chromedriver88'
         if operating_system == "Linux":
             architecture = platform.architecture()[0]
             if architecture == '32bit':
                 chrome_driver = 'linux_chromedriver65_32bit'
             else:
-                chrome_driver = 'linux_chromedriver86_64bit'
+                chrome_driver = 'linux_chromedriver87_64bit'
             os.chmod(os.path.join(drivers_dir, chrome_driver), 0o777)
         elif operating_system == "Windows":
-            chrome_driver = 'win_chromedriver86.exe'
+            chrome_driver = 'win_chromedriver88.exe'
 
         self.driver = selenium.webdriver.Chrome(
             options=options,
@@ -94,7 +95,8 @@ class Mexico:
 
     def __del__(self):
         self.driver.quit()
-        shutil.rmtree(self.downloads_dir)
+        if os.path.isdir(self.downloads_dir):
+            shutil.rmtree(self.downloads_dir)
 
     def __manual_click(self, element):
         WebDriverWait(self.driver, 10).until(
